sumarize pdouble and adouble
