+++ ADOL-C/include/adolc/ad_types.h
@@ -1,8 +1,22 @@
+/*----------------------------------------------------------------------------
+ ADOL-C -- Automatic Differentiation by Overloading in C++
+ File:     ad_types.h
+ Revision: $Id$
+ Contents: class for parameter dependent functions
+
+ Copyright (c) Kshitij Kulshreshtha
+
+ This file is part of ADOL-C. This software is provided as open source.
+ Any use, reproduction, or distribution of the software constitutes
+ recipient's acceptance of the terms of the accompanying license file.
+
+----------------------------------------------------------------------------*/
+
 /* ---------------------------------------------------------------------------
  ADOL-C -- Automatic Differentiation by Overloading in C++
 
  Revision: $Id$
- Contents: adouble.h contains the basis for the class of adouble
+ Contents: ad_types.h contains the basis for the class of adouble
            included here are all the possible functions defined on
            the adouble class.  See below for further explanation.
 
@@ -16,11 +30,11 @@
 
 ---------------------------------------------------------------------------*/
 
-#ifndef ADOLC_ADOUBLE_H
-#define ADOLC_ADOUBLE_H
+#ifndef ADOLC_AD_TYPES_H
+#define ADOLC_AD_TYPES_H
 
+#include "taping_p.h"
 #include <adolc/internal/common.h>
-#include <adolc/param.h>
 #include <cmath>
 #include <cstdio>
 #include <cstdlib>
@@ -36,6 +50,8 @@
 struct tape_location {
   size_t loc_;
 };
+
+class pdouble;
 
 /****************************************************************************/
 /*                                                            CLASS ADOUBLE */
@@ -56,7 +72,7 @@ public:
 
   adouble &operator=(const double coval);
   adouble &operator=(const adouble &a);
-  adouble &operator=(adouble &&a);
+  adouble &operator=(adouble &&a) noexcept;
   adouble &operator=(const pdouble &p);
 
   inline double getValue() const {
@@ -66,6 +82,7 @@ public:
   }
   inline double value() const { return getValue(); };
   inline size_t getLoc() const { return tape_loc_.loc_; };
+  inline size_t loc() const { return tape_loc_.loc_; };
 
   inline void setValue(const double coval) {
     ADOLC_OPENMP_THREAD_NUMBER;
@@ -90,10 +107,7 @@ public:
 
   adouble &operator/=(const double coval);
   adouble &operator/=(const adouble &a);
-  inline adouble &adouble::operator/=(const pdouble &p) {
-    *this *= recipr(p);
-    return *this;
-  }
+  adouble &operator/=(const pdouble &p);
 
   adouble operator++(int);
   adouble operator--(int);
@@ -114,8 +128,96 @@ private:
   int valid{1};
 };
 
+class ADOLC_DLL_EXPORT pdouble {
+
+public:
+  ~pdouble() = default;
+  pdouble(const pdouble &) = delete;
+  pdouble() = delete;
+  pdouble(pdouble &&) = delete;
+  pdouble &operator=(pdouble &&) = delete;
+
+  pdouble(const double pval);
+  explicit pdouble(tape_location tape_loc);
+
+  static pdouble mkparam(const double pval);
+  static inline pdouble getparam(size_t loc_) {
+    return pdouble{tape_location{loc_}};
+  };
+  explicit operator adouble() const;
+
+  inline size_t getLoc() const { return tape_loc_.loc_; }
+  inline double getValue() const {
+    ADOLC_OPENMP_THREAD_NUMBER;
+    ADOLC_OPENMP_GET_THREAD_NUMBER;
+    return ADOLC_GLOBAL_TAPE_VARS.pStore[tape_loc_.loc_];
+  }
+  inline void setValue(const double pval) {
+    ADOLC_OPENMP_THREAD_NUMBER;
+    ADOLC_OPENMP_GET_THREAD_NUMBER;
+    ADOLC_GLOBAL_TAPE_VARS.pStore[tape_loc_.loc_] = pval;
+  }
+
+private:
+  tape_location tape_loc_;
+};
+
 std::ostream &operator<<(std::ostream &, const adouble &);
 std::istream &operator>>(std::istream &, const adouble &);
+
+/*--------------------------------------------------------------------------*/
+/* sign operators */
+ADOLC_DLL_EXPORT adouble operator+(const adouble &a);
+ADOLC_DLL_EXPORT adouble operator+(adouble &&a);
+
+ADOLC_DLL_EXPORT adouble operator-(const adouble &a);
+ADOLC_DLL_EXPORT adouble operator-(adouble &&a);
+
+/*--------------------------------------------------------------------------*/
+/* binary operators */
+ADOLC_DLL_EXPORT adouble operator+(const adouble &a, const adouble &b);
+ADOLC_DLL_EXPORT adouble operator+(adouble &&a, const adouble &b);
+ADOLC_DLL_EXPORT adouble operator+(const adouble &a, adouble &&b);
+inline adouble operator+(adouble &&a, adouble &&b) { return std::move(a) + b; }
+
+ADOLC_DLL_EXPORT adouble operator+(const double coval, const adouble &a);
+ADOLC_DLL_EXPORT adouble operator+(const double coval, adouble &&a);
+
+adouble operator+(const adouble &a, const double coval);
+adouble operator+(adouble &&a, const double coval);
+
+ADOLC_DLL_EXPORT adouble operator-(const adouble &a, const adouble &b);
+ADOLC_DLL_EXPORT adouble operator-(adouble &&a, const adouble &b);
+ADOLC_DLL_EXPORT adouble operator-(const adouble &a, adouble &&b);
+inline adouble operator-(adouble &&a, adouble &&b) { return std::move(a) - b; }
+
+ADOLC_DLL_EXPORT adouble operator-(const double coval, const adouble &a);
+ADOLC_DLL_EXPORT adouble operator-(const double coval, adouble &&a);
+
+adouble operator-(const adouble &a, const double coval);
+adouble operator-(adouble &&a, const double coval);
+
+ADOLC_DLL_EXPORT adouble operator*(const adouble &a, const adouble &b);
+ADOLC_DLL_EXPORT adouble operator*(adouble &&a, const adouble &b);
+ADOLC_DLL_EXPORT adouble operator*(const adouble &a, adouble &&b);
+inline adouble operator*(adouble &&a, adouble &&b) { return std::move(a) * b; }
+
+ADOLC_DLL_EXPORT adouble operator*(const double coval, const adouble &a);
+ADOLC_DLL_EXPORT adouble operator*(const double coval, adouble &&a);
+
+adouble operator*(const adouble &a, const double coval);
+adouble operator*(adouble &&a, const double coval);
+
+ADOLC_DLL_EXPORT adouble operator/(const adouble &a, const adouble &b);
+ADOLC_DLL_EXPORT adouble operator/(adouble &&a, const adouble &b);
+ADOLC_DLL_EXPORT adouble operator/(const adouble &a, adouble &&b);
+inline adouble operator/(adouble &&a, adouble &&b) { return std::move(a) / b; }
+
+ADOLC_DLL_EXPORT adouble operator/(const double coval, const adouble &a);
+ADOLC_DLL_EXPORT adouble operator/(const double coval, adouble &&a);
+
+adouble operator/(const adouble &a, const double coval);
+adouble operator/(adouble &&a, const double coval);
 
 /* Comparison  */
 #ifdef ADOLC_ADVANCED_BRANCHING
@@ -184,60 +286,6 @@ bool operator<(const adouble &a, const double coval);
 
 bool operator>(const double coval, const adouble &a);
 bool operator>(const adouble &a, const double coval);
-
-/*--------------------------------------------------------------------------*/
-/* sign operators */
-ADOLC_DLL_EXPORT adouble operator+(const adouble &a);
-ADOLC_DLL_EXPORT adouble operator+(adouble &&a);
-
-ADOLC_DLL_EXPORT adouble operator-(const adouble &a);
-ADOLC_DLL_EXPORT adouble operator-(adouble &&a);
-
-/*--------------------------------------------------------------------------*/
-/* binary operators */
-ADOLC_DLL_EXPORT adouble operator+(const adouble &a, const adouble &b);
-ADOLC_DLL_EXPORT adouble operator+(adouble &&a, const adouble &b);
-ADOLC_DLL_EXPORT adouble operator+(const adouble &a, adouble &&b);
-inline adouble operator+(adouble &&a, adouble &&b) { return std::move(a) + b; }
-
-ADOLC_DLL_EXPORT adouble operator+(const double coval, const adouble &a);
-ADOLC_DLL_EXPORT adouble operator+(const double coval, adouble &&a);
-
-adouble operator+(const adouble &a, const double coval);
-adouble operator+(adouble &&a, const double coval);
-
-ADOLC_DLL_EXPORT adouble operator-(const adouble &a, const adouble &b);
-ADOLC_DLL_EXPORT adouble operator-(adouble &&a, const adouble &b);
-ADOLC_DLL_EXPORT adouble operator-(const adouble &a, adouble &&b);
-inline adouble operator-(adouble &&a, adouble &&b) { return std::move(a) - b; }
-
-ADOLC_DLL_EXPORT adouble operator-(const double coval, const adouble &a);
-ADOLC_DLL_EXPORT adouble operator-(const double coval, adouble &&a);
-
-adouble operator-(const adouble &a, const double coval);
-adouble operator-(adouble &&a, const double coval);
-
-ADOLC_DLL_EXPORT adouble operator*(const adouble &a, const adouble &b);
-ADOLC_DLL_EXPORT adouble operator*(adouble &&a, const adouble &b);
-ADOLC_DLL_EXPORT adouble operator*(const adouble &a, adouble &&b);
-inline adouble operator*(adouble &&a, adouble &&b) { return std::move(a) * b; }
-
-ADOLC_DLL_EXPORT adouble operator*(const double coval, const adouble &a);
-ADOLC_DLL_EXPORT adouble operator*(const double coval, adouble &&a);
-
-adouble operator*(const adouble &a, const double coval);
-adouble operator*(adouble &&a, const double coval);
-
-ADOLC_DLL_EXPORT adouble operator/(const adouble &a, const adouble &b);
-ADOLC_DLL_EXPORT adouble operator/(adouble &&a, const adouble &b);
-ADOLC_DLL_EXPORT adouble operator/(const adouble &a, adouble &&b);
-inline adouble operator/(adouble &&a, adouble &&b) { return std::move(a) / b; }
-
-ADOLC_DLL_EXPORT adouble operator/(const double coval, const adouble &a);
-ADOLC_DLL_EXPORT adouble operator/(const double coval, adouble &&a);
-
-adouble operator/(const adouble &a, const double coval);
-adouble operator/(adouble &&a, const double coval);
 
 /*--------------------------------------------------------------------------*/
 /* unary operators */
@@ -346,7 +394,7 @@ ADOLC_DLL_EXPORT adouble pow(const double base, const adouble &a);
 
 //-----------------------------------------------------------
 /* User defined version of logarithm to test extend_quad macro */
-ADOLC_DLL_EXPORT adouble myquad(double x);
+ADOLC_DLL_EXPORT double myquad(double x);
 
 /*---------------------------------------------------------------*/
 /*                                Conditional assignments */
@@ -378,6 +426,232 @@ void ADOLC_DLL_EXPORT adolc_vec_dot(adouble *const dest,
 void ADOLC_DLL_EXPORT adolc_vec_axpy(adouble *const res, const adouble &a,
                                      const adouble *const vec_a,
                                      const adouble *const vec_b, size_t size);
+
+adouble operator-(const pdouble &p);
+adouble operator+(const adouble &a, const pdouble &p);
+adouble operator+(adouble &&a, const pdouble &p);
+inline adouble operator+(const pdouble &p, const adouble &a) { return a + p; };
+inline adouble operator+(const pdouble &p, adouble &&a) {
+  return std::move(a) + p;
+}
+inline adouble operator+(const pdouble &p, const double coval) {
+  return coval + adouble(p);
+}
+inline adouble operator+(const double coval, const pdouble &p) {
+  return coval + adouble(p);
+}
+
+adouble operator-(const adouble &a, const pdouble &p);
+adouble operator-(adouble &&a, const pdouble &p);
+inline adouble operator-(const pdouble &p, const adouble &a) {
+  return (-a) + p;
+}
+inline adouble operator-(const pdouble &p, adouble &&a) {
+  return (-std::move(a)) + p;
+}
+inline adouble operator-(const pdouble &p, const double coval) {
+  return adouble(p) - coval;
+}
+inline adouble operator-(const double coval, const pdouble &p) {
+  return coval - adouble(p);
+}
+
+adouble operator*(const adouble &a, const pdouble &p);
+adouble operator*(adouble &&a, const pdouble &p);
+inline adouble operator*(const pdouble &p, const adouble &a) { return a * p; };
+inline adouble operator*(const pdouble &p, adouble &&a) {
+  return std::move(a) * p;
+}
+inline adouble operator*(const pdouble &p, const double coval) {
+  return coval * adouble(p);
+}
+inline adouble operator*(const double coval, const pdouble &p) {
+  return coval * adouble(p);
+}
+
+adouble recipr(const pdouble &p);
+adouble operator/(const pdouble &p, const adouble &a);
+adouble operator/(const pdouble &p, adouble &&a);
+inline adouble operator/(const adouble &a, const pdouble &p) {
+  return a * recipr(p);
+}
+inline adouble operator/(const pdouble &p, const double coval) {
+  return adouble(p) / coval;
+}
+inline adouble operator/(const double coval, const pdouble &p) {
+  return coval * recipr(p);
+}
+
+#ifdef ADOLC_ADVANCED_BRANCHING
+adouble operator!=(const adouble &a, const pdouble &p);
+adouble operator!=(adouble &&a, const pdouble &p);
+inline adouble operator!=(const pdouble &p, const adouble &a) {
+  return (a != p);
+}
+inline adouble operator!=(const pdouble &p, adouble &&a) {
+  return (std::move(a) != p);
+}
+
+adouble operator==(const adouble &a, const pdouble &p);
+adouble operator==(adouble &&a, const pdouble &p);
+inline adouble operator==(const pdouble &p, const adouble &a) {
+  return (a == p);
+}
+inline adouble operator==(const pdouble &p, adouble &&a) {
+  return (std::move(a) == p);
+}
+
+adouble operator<=(const adouble &a, const pdouble &p);
+adouble operator<=(adouble &&a, const pdouble &p);
+inline adouble operator<=(const pdouble &p, const adouble &a) {
+  return (a >= p);
+}
+inline adouble operator<=(const pdouble &p, adouble &&a) {
+  return (std::move(a) >= p);
+}
+
+adouble operator>=(const adouble &a, const pdouble &p);
+adouble operator>=(adouble &&a, const pdouble &p);
+inline adouble operator>=(const pdouble &p, const adouble &a) {
+  return (a <= p);
+}
+inline adouble operator>=(const pdouble &p, adouble &&a) {
+  return (std::move(a) <= p);
+}
+
+adouble operator>(const adouble &a, const pdouble &p);
+adouble operator>(adouble &&a, const pdouble &p);
+inline adouble operator>(const pdouble &p, const adouble &a) { return (a < p); }
+inline adouble operator>(const pdouble &p, adouble &&a) {
+  return (std::move(a) < p);
+}
+
+adouble operator<(const adouble &a, const pdouble &p);
+adouble operator<(adouble &&a, const pdouble &p);
+inline adouble operator<(const pdouble &p, const adouble &a) { return (a > p); }
+inline adouble operator<(const pdouble &p, adouble &&a) {
+  return (std::move(a) > p);
+}
+
+#else  // ADOLC_ADVANCED_BRANCHING
+
+inline bool operator!=(const adouble &a, const pdouble &p) {
+  return ((a - p) != 0);
+}
+inline bool operator!=(const pdouble &p, const adouble &a) { return (a != p); }
+inline bool operator!=(const pdouble &p, adouble &&a) {
+  return (std::move(a) != p);
+}
+
+inline bool operator==(const adouble &a, const pdouble &p) {
+  return ((a - p) == 0);
+}
+inline bool operator==(const pdouble &p, const adouble &a) { return (a == p); }
+inline bool operator==(const pdouble &p, adouble &&a) {
+  return (std::move(a) == p);
+}
+
+inline bool operator<=(const adouble &a, const pdouble &p) {
+  return ((a - p) <= 0);
+}
+inline bool operator>=(const adouble &a, const pdouble &p) {
+  return ((a - p) >= 0);
+}
+inline bool operator<=(const pdouble &p, const adouble &a) { return (a >= p); }
+inline bool operator<=(const pdouble &p, adouble &&a) {
+  return (std::move(a) >= p);
+}
+inline bool operator>=(const pdouble &p, const adouble &a) { return (a <= p); }
+inline bool operator>=(const pdouble &p, adouble &&a) {
+  return (std::move(a) <= p);
+}
+
+inline bool operator>(const adouble &a, const pdouble &p) {
+  return ((a - p) > 0);
+}
+inline bool operator<(const adouble &a, const pdouble &p) {
+  return ((a - p) < 0);
+}
+
+inline bool operator>(const pdouble &p, const adouble &a) { return (a < p); }
+inline bool operator>(const pdouble &p, adouble &&a) {
+  return (std::move(a) < p);
+}
+
+inline bool operator<(const pdouble &p, const adouble &a) { return (a > p); }
+inline bool operator<(const pdouble &p, adouble &&a) {
+  return (std::move(a) > p);
+}
+#endif // ADOLC_ADVANCED_BRANCHING
+
+/* unary operators */
+inline adouble exp(const pdouble &p) { return exp(adouble(p)); }
+inline adouble log(const pdouble &p) { return log(adouble(p)); }
+inline adouble log10(const pdouble &p) { return log10(adouble(p)); }
+inline adouble sqrt(const pdouble &p) { return sqrt(adouble(p)); }
+inline adouble cbrt(const pdouble &p) { return cbrt(adouble(p)); }
+inline adouble sin(const pdouble &p) { return sin(adouble(p)); }
+inline adouble cos(const pdouble &p) { return cos(adouble(p)); }
+inline adouble tan(const pdouble &p) { return tan(adouble(p)); }
+inline adouble asin(const pdouble &p) { return asin(adouble(p)); }
+inline adouble acos(const pdouble &p) { return acos(adouble(p)); }
+inline adouble atan(const pdouble &p) { return atan(adouble(p)); }
+inline adouble sinh(const pdouble &p) { return sinh(adouble(p)); }
+inline adouble cosh(const pdouble &p) { return cosh(adouble(p)); }
+inline adouble tanh(const pdouble &p) { return tanh(adouble(p)); }
+inline adouble asinh(const pdouble &p) { return asinh(adouble(p)); }
+inline adouble acosh(const pdouble &p) { return acosh(adouble(p)); }
+inline adouble atanh(const pdouble &p) { return atanh(adouble(p)); }
+inline adouble erf(const pdouble &p) { return erf(adouble(p)); }
+inline adouble erfc(const pdouble &p) { return erfc(adouble(p)); }
+inline adouble fabs(const pdouble &p) { return fabs(adouble(p)); }
+inline adouble ceil(const pdouble &p) { return ceil(adouble(p)); }
+inline adouble floor(const pdouble &p) { return floor(adouble(p)); }
+
+inline adouble fmin(const adouble &a, const pdouble &p) {
+  return fmin(a, adouble(p));
+}
+inline adouble fmin(const pdouble &p, const adouble &a) { return fmin(p, a); }
+inline adouble fmin(const pdouble &p, const pdouble &q) {
+  return fmin(p, adouble(q));
+}
+inline adouble fmin(const double coval, const pdouble &p) {
+  return fmin(coval, adouble(p));
+}
+inline adouble fmin(const pdouble &p, const double coval) {
+  return fmin(adouble(p), coval);
+}
+
+inline adouble fmax(const adouble &a, const pdouble &p) {
+  return fmax(a, adouble(p));
+}
+inline adouble fmax(const pdouble &p, const adouble &a) { return fmax(a, p); }
+inline adouble fmax(const pdouble &p, const pdouble &q) {
+  return fmax(p, adouble(q));
+}
+inline adouble fmax(const double coval, const pdouble &p) {
+  return fmax(coval, adouble(p));
+}
+inline adouble fmax(const pdouble &p, const double coval) {
+  return fmax(adouble(p), coval);
+}
+
+inline adouble ldexp(const pdouble &p, const int exp) {
+  return ldexp(adouble(p), exp);
+}
+inline adouble frexp(const pdouble &p, int *exp) {
+  return frexp(adouble(p), exp);
+}
+
+adouble pow(const adouble &a, const pdouble &p);
+adouble pow(adouble &&a, const pdouble &p);
+adouble pow(const pdouble &p, const adouble &a);
+inline adouble pow(const pdouble &p, const double coval) {
+  return pow(adouble(p), coval);
+}
+inline adouble pow(const double coval, const pdouble &p) {
+  return pow(coval, adouble(p));
+}
 
 /* numeric_limits<adouble> specialization
  *
@@ -460,4 +734,4 @@ template <> struct std::numeric_limits<adouble> {
       std::numeric_limits<double>::round_style;
 };
 
-#endif /* ADOLC_ADOUBLE_H */
+#endif // ADOLC_AD_TYPES_H
